[MDC-1343] twamp: implement -k<key> for twamp-light mode
Add -k <key> to the cli, to supply an authentication/authorization key
(in base64 encoding).

Currently, implement it only for TWAMP-light mode, as a SIMET2-specific
extension inside the padding area, using a STAMP Private Extension
(RFC-8972) at a specific payload location.

For full TWAMP mode, -k is ignored for now.

This authentication/authorization key will be used for the authenticated
mode shared secret if we ever implement that, at which point the SIMET2
extension can be deprecated.

The STAMP Private Extension requires a IANA-assigned Private Enterprise
Number ("PEN").  The IANA-assigned PEN for CEPTRO.br is 60267.