[MDC-1343] twamp: prefix report_* identifiers with lmap_
We will add a second report type (summary report), which is unrelated to
the LMAP report.  Rename indentifiers and fields accordingly, so that we
know which ones are related to lmap.

Rename struct twamp_parameters.report_mode to lmap_report_mode.
Rename struct twamp_report_private.root to lmap_root.