[MDC-1343] twamp: rework LMAP report output
Add a "none" LMAP report mode, that supresses the *generation* of the
LMAP report entirely, and change the implementation of the "-o" option
to redirect just the LMAP report serialization output, instead of
redirecting stdout itself.

This mode is requested with CLI option "-r2".

It is possible to set the -o option to "" (empty) or "-", which will
cause the report to go to stdout.

Refactor the report serialization into a separate function, as we will
reuse it for the summary report later, and defer opening the LMAP report
output stream so that we do not needlessly create/truncate it should the
program exit early for whatever reason, or when no LMAP report is to be
generated.

Fix twamp_report_render_lmap() to actually preserve the json object tree
in the twamp_private_report struct in the case where we are not
extending a previously existing report with other components.  This is
not the normal program flow right now, though.

After these changes, twamp_report_render_lmap() exists early without
doing any work or generating JSON structures in memory when the LMAP
report mode is "none".

Should twamp_parameters.twamp_report_path be set, a new output stream is
created, written to, and closed.  If twamp_parameters.twamp_report_path
is NULL and twamp_report_output is not NULL, the twamp_report_output
stream is written to (and _not_ closed).  When both are NULL, the report
is not serialized.