[MDC-1350] simet-runner: add optinal peer reordering map (v2)
Add a reordering map layer to the discover_init()/discover_next_peer()
API.

When GLOBAL_STATE_PEER_IDXMAP is unset (the default), we use the
service-list output as-is.

When GLOBAL_STATE_PEER_IDXMAP is set with a space-separated list of
indexes (starts at zero, cannot be negative), we use the service-list
entries in those indexes in map order, until one of them succeeds
[in allowing a measurement to be taken].

The last entry in the map must be a negative number or we will loop
trying the last entry until it succeeds (possibly forever).