[MDC-1350] simet-runner: limit wait for server-monitor reply
Limit attempts to talk to each location's server-monitor to 15s for TCP
connect, and 30s for API reply.

Without preflight tests, simet-runner often attempts to connect to
locations that do not actually have a viable network path, and thus
cURL's timeout settings are very relevant, and the defaults are too
long.

In fact, 15s for TCP connect might be too much (it cannot be less than
5s, though).  It can be further reduced later.