[MDC-1350] simet-runner: latency-based server selection parameters
Call the service-list API in background with updated timeout and cURL
parameters, and add a call to the twampquick server selection API, also
in background, to get the required parameters to run a latency-based
server selection algorithm.

This lets the two API calls resolve at the same time, to not add further
latency to the beginning of the measurement.

The results of the twampquick server-selection API call are stored to
twampquick_parameters.json, but not used for anything yet.  If the call
fails, the file is not created.

Also, add the endpoint for the twampquick server-selection API call to
simet-ma.conf.