[MDC-1350] simet-runner: latency-based server selection
Walk the services list, and run a twamp-quick measurement on any
non-public peers.  These measurements are run in parallel.

Order the resulting set of twamp-quick measurements by increasing
latency, skipping any peers that failed to respond.

Append to the final result set, the public (fallback) peers.

If any of that fails, use the full services list with the original
ordering.