[MDC-1350] simet-runner: respect localityListIndex when ordering
Although we are allowed (for now!) to reorder across localityListIndex
boundaries, we still have to respect it as a tie breaker.

Change the index metadata to be a colon separated field, and use colon
separated fields for sorting (in "numeric mode") explicitly in the POSIX
"sort" call.

We should have the server-list array position as the last element in
that field (it is the result of the sorting, for the ordered map), and
add other sort keys before it (such as localityListIndex).