[MDC-1350] simet-runner: serialize server selection on low RAM
We can't afford to run tens of copies of twampc in parallel on
small-memory devices such as some SIMETBoxes.

Refactor the server selection a bit so that it will run the server
selection measurements serially when GLOBAL_SERIALIZE_SERVERSEL is set
to 1.

Set GLOBAL_SERIALIZE_SERVERSEL to 1 when MemAvailable (in /proc/meminfo)
is below 25000kB.  This should detect 32MiB-RAM SIMETBoxes as well as
anything else that is severely overloaded.

Note that we still run two twampc in parallel: one for IPv4, one for
IPv6.