[MDC-1343] twamp: rework -R and -r report control options
Rework the recently changed CLI interface into something more
future-proof:

* Remove the recently added "-r2" mode
* Change -R to take a list of report types to enable

Report type "lmap" (or "LMAP") is the default, and outputs the LMAP
report according to the "-r0" or "-r1" options to stdout, or to the
destination specified by the "-o" option.

non-LMAP reports output to the destination specified by the recently
added "-O" option.

If neither of the "-O" or "-o" options are used, and there are both LMAP
and non-LMAP reports enable, one ends up with JSON objects back-to-back
right now, which is hard to parse.  This might be changed in the future.

Output order to stdout is fixed: LMAP (if enabled), then non-LMAP (if
enabled).

To disable all reporting, use -R "".  To disable LMAP reporting, it is
enough to not mention "lmap" (or "LMAP") in -R.

Typical ways to call twampc are, now:

  -o lmap.json -O summary.json -R lmap,summary
  -R summary
  -R lmap