[MDC-1350] simet-runner: add workaround for limited busybox sort
When busybox is compiled without enabling "big" sort, it supports only
-n -u and -r, and ignores -t and -k.

Add (hopefully) enough leading zeroes to all numeric keys to cover the
expected range, so they will sort correctly.  Do this only right before
the sort, and remove it right after: leading zeroes cause the shell to
interpret numeric values as octal.

This is ugly, but it works.  Too bad we can't depend on bash, GNU sed
and non-minimal "sort", it would make the code a *lot* simpler.