source: modernize C support in autotools
Switch from AC_PROG_CC_STDC to AC_PROG_CC, as what counts as the current
version of STDC is not really an useful definition when old autoconf is
used...

We want C11 with GNU extensions, so change that to AC_PROG_CC + compiler
checks for -std=gnu11 or -std=c11.  Also, fix the way we handle CFLAGS
in configure.ac to not override whatever autoconf macros detected as
required.

While at it, remove some currently unnecessary header or function checks
that are just slowing down the build.