source: unbreak openwrt 15 build
OpenWRT 15.05 has new-enough gcc (4.6/4.8) with enough C11 support for
our needs, but its libc does not, and assert.h does not define
static_assert().

Since C23 decided to do away with _Static_assert(), manually define
static_assert as an alias to _Static_assert if it is missing and the
compiler doesn't announce a __STDC_VERSION__ that could be new enough
for C23 (since in C23 it will be a keyword, and might not be seen by #if
defined() that supposedly looks only for macros).