[MDC-199] twamp: avoid early exit on receive thread
When comparing timespecs with timespec_lt(), they *must* be normalized
(as documented!).  Unfortunately, ts_stop was not being normalized after
addition.

Due to the lack of normalization, should ts_stop.tv_nsec be >=
1000000000, timespec_lt() would misbehave and return a false positive.

This resulted in a (silent!) early termination of the receive thread.
All remaining packets would be considered lost when this issue triggers.

Fix it by normalizing ts_stop properly in twamp_callback_thread().

Fixes: commit f845eebf21c98c62d5dce4f9ab0d6da1229cb5e8
       [MDC-199] twamp: get rid of gettimeofday() and some warnings