[MDC-1350] simet-runner: change server-selection-by-RTT ordering
When doing RTT-based server selection, first sort by the service list
tier/group, and then sort by rtt.

To do that, refactor the subtask_serversel component a bit, so that we
keep three arrays: RTT, sort data (DAT), and server-list index/peer id
(IDX).

The service-list will be changed to better aggregate peers in tiers so
that we prefer ASN-internal and directy-connected IXPs in the
appropriate order.