[MDC-1350] simet-runner: update ServiceList report metric
Add the GLOBAL_STATE_PEER_IDXMAP map with the latency-based server
selection result to the ServiceList debugging report metric.

This required some refactoring of templates.sh::task_json_template(),
to allow multiple columns from multiple JSON files.

During the refactoring, a bug in task_json_template() was found, and
fixed: it would not handle properly any LFs in the JSON, and would
generate multiple rows (with JSON fragments).  This will happen if the
Service List API ever decides to send formatted JSON output with line
breaks, but there is no real impact, since it is just debugging
information, and the end result should still be syntatically valid and
accepted as a measurement report.