[MDC-1358] simet-runner: do not drop peer due to ipv6 loss
Do not drop a peer because IPv6 had 100% loss, this is usually related
to no IPv6 connectivity at all, instead of a problem to reach a specific
peer.

This is needed because, among other failure scenarios, twampc won't get
a send error when it has a non-global ULA address, but no global IPv6
connectivity.  And that is a rather common situation.

Note that we might still drop entirely a location that has high IPv6
loss (as long as it is not 100%).  This is on purpose.