tcpbw: ensure proper struct alignment
report.h defines an "opaque struct" that has private fields defined only
in report.c.  That would've been fine, if not for the fact that structs
can have different minimum alignment requirements depending on their
members.

Add a fake pointer to the exposed struct, to force the required minimum
alignment and get gcc to stop warning about it.