[MDC-788] tcpbw: early-free JSON report data structures
libjson-c data structures can get quite large, and they hold both the
native data structures as well as the serialized/rendered JSON text.

We'll make that problem even worse with tcp_info JSON metadata, so free
them ASAP.

In the future, we should address this by vendoring up-to-date json-c
(and adding automake glue to use the vendored one if the system version
is too old or not found).  That would enable the use of newer json-c
features for no-strdup() of static RO data, etc.