[MDC-720] deb: be far more assertive about automated updates
Ignore the local administrator settings re. unattended upgrades and have
the cron.weekly script enforce an apt-get install of the SIMET packages.
This forces their upgrade if they are outdated.  There is a guard that
ensures it is only done if the main package is not in a removed state.

We've seen way too many systems where just depending on unattended-
upgrades is *not* enough: they just won't update.  Likely, these systems
already had unattended-upgrades installed, but disabled.

Note that we cannot just force-enable unattended-upgrades, as it updates
the whole system and not just our packages.

While at it, add an "apt-get -f install" run when things fail, to
hopefully fix pending brokenness and give an eventual next attempt a
chance to work.

The cron script guards require that /opt/simet/bin/simet_register_ma.sh
exist.  If we move or rename that script, the guards need to be updated
to match otherwise it would disable the cron scripts.  Add a self-test
to Debian rules that will abort the package build if we screw that up.