[MDC-1363] simet-api: add QRcode and /etc/issue generators
Add the simet_qrcode.sh script, to generate a QRcode with either a
pairing URL, or invariant identification information (such as the main
MAC address, and the virtual label).

Add a script that can be called by system glue to generate console
banners.  Currently, we do not add a QRcode as the app that might use it
is not yet available.

Note: QR codes are generated only if "qrencode" is installed (and we
only "suggest" it for now in the packaging).