[MDC-848] simet-api: remove fping-to-ixbr measurement
Disable and remove the client side of the fping-to-ixbr measurement.

The client-side of the fping-to-ixbr measurement should have been
removed a long time ago, actually, as the team decided the error margin
of ICMP ECHO/REPLY was large enough to make it hard to derive any real
insight from the measurement results.

The server-side processing of this measurement's results has been
disabled a long time ago.

We retain the LMAP capability of doing fping measurements on demand, but
this might be removed as well in the future.  Nothing makes use of this
capability right now.