[MDC-1368] debian: create vlabel on initscript, not postinst
In the presence of vendor hooks or regular hooks provided by other
packages, calling simet_create_vlabel.sh on postinst is a hassle, as it
requires package ordering to ensure the hooks are in-place before the
postinst is called.

Also, now simet_create_vlabel.sh has to support temporary failures, so
it is wise to be able to retry vlabel creation.

Since it is not a problem to call it when a vlabel is already present
(it will decay to the equivalent of a simet_read_vlabel.sh call), just
call it on initscript start (but not restart/reload), instead of
calling it on postinst.