[MDC-1377] tcpbw: tcp MSS awareness
Track the maximum MSS of the measurement streams (clamped to 64KiB),
warn if anything is amiss.  Warn, and assume a MSS of 1400 bytes if we
fail to detect it (this should not happen).

Enable TCP_CORK mode: we want to get the kernel to generate MSS-sized
segments as much as possible.  The kernel might still send (read: will
eventually send) shorter segments for several reasons, so we don't
bother forcing send() to always use a multiple of MSS.

Rewind buffer a bit earlier (2 * MSS), and abort the measurement if our
application buffer is not at least ten times the MSS -- otherwise we'd
be too much BDP-constrained.