tcpbw: early exit if peer closes every stream
If the MP closes all measurement streams with more than two seconds left
until the end of the measurement, abort with an error.

This way, on a network or MP error condition where all measurement
streams get a TCP RST or an early EOF, we avoid waiting for the entire
upload and/or download measurement time window to stop.

This change also invalidates low-quality partial/short measurements.