[MDC-1377] tcpbw: RTT awareness: retrieve minimum RTT from TCP_INFO
Request from the kernel the iRTT or sRTT for the TCP streams after we
have sent the stream header data (so, a bit better than just the iRTT),
and store it for later use.

Since TCP_INFO also gives us the MSS, refactor the MSS retrieval to get
it from the tcp_info structure, but fallback to TCP_MAXSEG if that
fails.