[MDC-1282] simet-api: detect pairing status without a vlabel
Just because the MA does not have a virtual label, or because the
virtual label has been explicitly disabled, doesn't mean the MA is
unpaired.

It just means it will be unavailable to pairing *if it is not yet
paired*.

Correctly update in-MA metadata with pairing information if it is
paired, regardless of the presence of a virtual label.