[MDC-1380] simet-api: lastrun timestamp for simet_register_ma.sh
It would be useful to be able to call simet_register_ma from events with
an unknown frequency and possibly bursty nature, such as network-is-up
events.

Overload the --boot mode to do just that: use the first line of the
lockfile to hold an UTC timestamp (from "date +%s -u"), and refuse to
run if we've already run sucessfully in the last 24H.

This handle cases such as a deny-listed MA that has been ordered to
erase its registration information.

Since we want to register again on package upgrades/downgrades, but
something weird might be going on (e.g. restart after package update,
instead of stop ; update package ; start), we don't remove the lockfile
but instead overwrite its contents (if present) with "0\n", which will
always trigger a new simet_register_ma.sh run.