[MDC-1395] tcpbw: abort measurement on measurement connection fail
We used to try to measure even if we did not get all measurement
connections we wanted.

This is being changed: on larger RTTs we easily hit per-TCP-stream
limits, and we need at least two or three extra streams to try to
compensate for packet loss / reordering events.  So we really don't want
to run a measurement with a reduced number of streams.

Abort TCPBW measurement with an error status should we fail to open the
desired number of measurement connections.  The typical reasons we end
up with less connections than we tried to open are: high packet loss or
extreme RTTs causing the 3-way handshake to get too near the TCP connect
timeout limit, and end-to-end connectivity loss with the server.