[MDC-1399] deb: increase system socket buffer size limits
For bandwidth measurements, we need TCP socket buffers that can grow to
match at least 25% of the BDP (for the typical measurement with 5 or 6
streams).

The system's socket buffer limit defaults are not good enough for modern
broadband service, even at modest 15-35ms RTTs: it is necessary to
increase them.

Unfortunately, it is not enough to just increase the maximum TCP socket
buffer size: increasing the default (initial) TCP socket buffer size
resulted in increased throughput on short-lived elephant flows (10s, all
bandwidth available to TCP), so do that as well.