[MDC-873][MDC-1380] simet-api: simet_register_ma: set umask to 0002
We want group-writeable MA registration files. simet_register_ma.sh
might run either as root (undesireable) or by a less privileged user,
and the group permission is used to enable this.

Set umask to 0002, instead of 0022.  And when moving $OUTFILE around
(which is created mode 0700 by mktemp), chmod it to og+rw, instead of
og+r.

Fixes: commit dca7bf6c084982b69f7bb1e62e4dd5171ee9832d
       [MDC-873] simet-api: simet_register_ma: ensure sane umask