[MDC-1402] simet-runner: use simet_lib.sh
Load in simet_lib.sh from defines.sh, and use simet_lib to unbreak the
fallback path of task_environment.

This allows us to remove lsb_agentinfo.sh and simetbox_agentinfo.sh.

Keep using the lmap helper for task_environment for consistency, and the
fallback path is still scheduled to be removed in the future.