[MDC-1380][MDC-1401] deb: provide hooks for dhcpdc (dhclient), pppd
Add hook for dhclient (dhcpdc) and pppd with the same functions of the
already provided ones for NetworkManager, networkd-dispatcherd and ifup.

This is important when dhcpdc.service or pppd is being used *instead* of
one of the other three network configuration methods, as opposed to
being called thorough ifupdown, etc.

Fixes: commit df7ac1541b55b707c505753e86d563f4a587529e
       ([MDC-1401] deb: start measurement engine on network hooks)
Fixes: commit e8a090e86f9e0d874caea064338e4120ddf496dd
       ([MDC-1380] deb: hook simet-register-ma to network changes)