[MDC-1410] simet-api: add simet-api/utils/ and simet_gen_vlabel
Add simet-api/utils as a project directory for smallish C programs.

This really asks for a refactoring of the whole mess with the other C
programs, using e.g. automake Include, and eliminating duplicated source
code through libs.  But for now, do it the hard (and messy way) by
duplicating everything needed yet one more time...

Add the simet_gen_vlabel utility, to generate N0-style vlabels from two
secrets (e.g. serial number and MAC address, invariant-FLASH dumps and
serial number, etc).  One secret goes in the command line (text), the
other goes in stdin (binary).  The two are *not* interchangeable:
swapping them will change the resulting virtual label.

base64.[ch] is a duplicate from the code in twamp-client-c/src/

sha256.[ch] is an adaptation of the code used by git, which is itself, a
very close adaptation of the reference code for SHA2 in the standards.

crc16.[ch] is an adaptiation of public domain code.  It is somewhat
optimized for speed and likely larger because of that, so it should be
replaced with something optimized for size in the future.