[MDC-1410] simet-api/utils: sha256: refactor HMAC_SHA256 for PBKDF2
In preparation for a much faster PBKDF2 implementation than the
reference one, duplicate HMAC_SHA256_CTX.hctx (SHA256_CTX) into inner
and outer contextes, as required by the enhanced PBKDF2 algorithm.

This allows us to move k_opad[] out of the HMAC context, which mitigates
the increase in size of HMAC_SHA256_CTX a bit.

Note that this separation of the inner and outer hash contextes is
useless for plain HMAC_SHA256.  Whether it is detrimental or not due to
cache effects or extra memory usage was not verified, as the improved
PBKDF2 that needs it is far more relevant at this time.