[MDC-1410] simet-api/utils: use PBKDF2-HMAC-SHA256 for N0 labels
Use PBKDF2-HMAC-SHA256 instead of HMAC-SHA256 when generating N0 labels,
to make it slightly harder to brute-force.

Add two runtime self-tests for PBKDF2-HMAC-SHA256 to simet_gen_vlabel,
they're fast enough, and we don't trust this code and the cross-
compilers that much yet.

Add -l and -n options to simet_gen_vlabel, for generic key length and
"parameter n".  For the N0 format, the key length must be from 128 to
256, and the "n" parameter is the iteration count, which must be at
least 1000.