[MDC-1410] simet-api/utils: base64: add no-padding and improve unit test
Make padding optional when encoding and decoding.  This adds an extra
parameter to base64_encode() and base64safe_encode(), so change call
sites accordingly.

When decoding, we have not added a "strict mode" that requires padding,
so padding became optional for now.

Update the unit tests to cover a lot of the no-padding cases.

Add base64_test to Automake's check_PROGRAMS.  We don't call it yet,
though, but at least now it is easy to build ("make check" from inside
simet-api or simet-api/utils), and it can be run manually when needed.