[MDC-1399] deb: don't write to /proc unless needed
If we know we'd write the same thing we have just read from
/proc/.../tcp_*mem, skip the write.

This avoids errors during reload of the simet-ma service, if the reload
was triggered from the nicbr-simet user (e.g. due to agent-id changes).

We can't just skip the whole thing if "test -w" fails: we want to warn
if that means the values in /proc are lower than what we wanted.

And it is /proc, so it is best to not depend on test -w, but rather
attempt the write and complain if it fails.

While at it, change do_update_proc_triplet() to run in a subshell, since
it messes with IFS.

Fixes: commit f3ddb2f68484cfecbdb202bf730454dc1be61f8a
       ([MDC-1399] deb: increase system socket buffer size limits)