[MDC-1380] deb: postinst: always reset AGENT_LOCK
Reset the contents of AGENT_LOCK even if LMAP_SCHEDULE_FILE handling
failed for some reason.

While at it, add explicit checks for LMAP_SCHEDULE_FILE and AGENT_LOCK
being defined and non-empty.  This is not fixing any bug: the implicit
failure of test -s or ln -sf when the target is "" already covered this
situation, but it is better to be explicit about these things.

Fixes: commit 700c0840100c70ac30ce0a654fc92645429edeb1
       ([MDC-1380] simet-api: lastrun timestamp for simet_register_ma.sh)