deb: simet-ma.init: safer handling of find command
Always use 'find "$VAR" -maxdepth 0 -type f' when we expect VAR to point
to a file, so that "find" will failsafe to doing nothing should VAR
point to a directory or should it be empty.

Fix some "find" calls so that the proper ordering of "-maxdepth" before
"-type" is used, otherwise the "find" command may fail.