[MDC-1419] simet-api: add lmap-empty-schedule.json template
We need "default" schedules for the extra lmapd instances, which should
be used when they are to sit around in idle mode.

The empty schedule will be used through symlinks, just like the
emergency schedule is used (by the main lmapd instance).

The initscripts (and possibly other glue, such as cron scripts) and
simet*_lmap-fetch-schedule.sh will need it.