[MDC-1425] inetup: rename confusing local variable
In simet_uptime2_recvmsg(), a variable named "processed" was used to
mean a frame matched a handling rule (i.e. that a message handler
was found for that message).

This does not mean the message has been processed.  It might not have
been processed yet, and in fact there are bugs in that logic that will
be fixed in the next commit.

Rename the "processed" variable to "handler_found".