[MDC-1425] inetup: properly handle partial messages
We might have incomplete messages in the socket receive buffer since TCP
is stream-oriented and has no concept of message boundaries unlike UDP.
The tcpaq layer allows us to handle this properly, but that does mean we
need to retry when tcpaq_peek* and tcpaq_discard returns 0 indicating
that we need to wait for more data to arrive in the scoket buffer.

Unfortunately, there is a bug in simet_uptime2_recvmsg():

* tcpaq_peek_nowait might return zero (incomplete object, must wait for
  TCP to receive more bytes)
* simet_uptime2_recvmsg() skips calling the message handler (correctly)
  in that case
* simet_uptime2_recvmsg() calls tcpaq_discard() unconditionally, which
  is *wrong*, and might even succeed in dropping the (not-yet-handled!)
  message as more data might have arrived in the socket buffer at that
  point.

Move the buggy tcpaq_discard() call to the code path that runs the
message handler, so that it will only execute when we did peek
(received) the full message.  Note that we discard regardless of
the message handler's result (but that result *is* returned to the
caller).