[MDC-1428] tcpaq: avoid tcpaq_discard() and tcpaq_drain returning EFAULT
It is always incorrect to call recv() with a NULL buffer and non-zero
buffer size.  recv() for TCP sockets with MSG_TRUNC under Linux don't
change that rule.  In fact, the Linux kernel will return EFAULT in that
case, just as it would without MSG_TRUNC.

Note that when buffer overflow checks are enabled, [some versions of ?]
GCC and glibc will map the zero-page, and the kernel will *not* return
EFAULT, hiding the issue.

Also, almost always the object being tcpaq_discard()'ed is already fully
in in_queue.buffer, which also sidesteps the issue, as no recv() is
done.

Rework tcpaq_discard() and tcpaq_drain() to use in_queue.buffer for
recv().

Note that tcpaq_drain() needs to be run in a loop until it returns zero
(or an error) in order to fully drain the socket queue, and this is not
changed by this patch.