[MDC-1430] deb: do install emergency schedule on install/upgrade
simet-ma.postinst was overwriting the emergency schedule with the empty
schedule, which results in a idle main lmap instance, unless it receives
a valid schedule from the LMAP controller at simet-lmapd service start.

Fix it so that we correcly switch to the emergency schedule, as that
gets replaced by a local schedule *or* a valid LMAP-controller supplied
schedule at simet-lmapd service start.

Fixes: commit 043c36d99296822192b89f6e5fd3b3b94061ce9a
       [MDC-1419] deb: simet-ma.postinst: handle multiple lmapd instances