[MDC-1427] sspoofer: fix LMAP report mode
CLI option -r (LMAP report mode) was inverted, resulting in invalid
reports when run from simet-runner.

Refactor the report mode handling to something closer to the one in
twamp-client-c, which is more advanced and uses enums, a style that is
far less prone to this kind of bug.

Adjust the lmapd configuration to run the sspooferc tasks in report mode
1, as expected by simet-lmapd.