[MDC-1434] twamp: refactor per-packet data storage
Move the per-packet RTT calculation to the receive loop, and store only
the relevant packet data instead of the whole packet.

If the timings derivated from the in-packet timestamps are strange, the
RTT is clamped to zero.  We flag this on the analysis loop and redo the
RTT calculation, to check whether it was actually zero or due to invalid
timings.

This refactoring will be needed for the UI callback support.