[MDC-1434] twamp: add UI callbacks for partial status
Add support for callbacks to show TWAMP measurement progress.  They are
not used on the CLI UI, since it just disturbs the measurement for no
good reason.

The UI *must* use either the callback or the results_summary JSON report
to show the final results when the measurement ends.

To use, define the preprocessor symbols for the callbacks in an include
before their definition in twamp.h (not yet present in this patch).
When the symbols are undefined, the header will provide suitable stubs
that the compiler optimizes away while still preserving type checking.