[MDC-1427] spoofer: ensure len returned by getsockname/getpeername fits
getsockname() and getpeername() may return an addrlen that is bigger
than the provided buffer (original value of addrlen), if the buffer was
too small.

We can either trim the returned addrlen to the real buffer size and
proceed with truncated, invalid information... or we can error-out in
that case (which is what this change does).

What we should not do, is to proceed with addrlen bigger than the real
buffer as the code was doing.  THAT is a buffer overflow waiting to
happen...

Note that this is *not* fixing an existing buffer overflow security bug,
because the code currently always uses buffers that are at least as
large as sizeof(sockaddr_storage).

But future changes to use smaller sockaddr buffers *would* introduce
such a security bug without this fix.