[MDC-1427] spoofer: fix minor memory leak in sspoofer_msghdl_msmtreq()
Fix a minor memory leak in sspoofer_msghdl_msmtreq().  While at it, give
"prefix_str" a better name since it is used to parse other values *and*
it is non-persistent.

Right now, this leak is unimportant because we just process a single
MSMTREQ message, but that might not remain true for long.