[MDC-1427] spoofer: free memory when exiting main loop
Free the server_pollfds array when we don't need it anymore, which might
make a little more memory available to sspoof_render_report().

Free the server and server cluster structures when we don't need them
anymore.  The program is about to exit at that point, so this is just to
remove distracting cruft from valgrind output.

The point of these changes is to make it easier to use valgrind to find
real memory leaks *inside* the main loop, therefore the error paths that
terminate the program early don't matter.