[MDC-1427] spoofer: fix minor memory leak in sspoofserver_destroy()
Free the measurement context structures in sspoof_server::msmt_queue and
sspoof_server::msmt_done, as well as sspoof_server::sid.str when
destroying the structure.

This code path only runs when the main loop exits right now, but fixing
the leak is useful both to ensure it wouldn't become a real leak in the
future, and to clear up valgrind output.