[MDC-1382] tcpbw: enforce a minimum stream start delay when enabled
When stream start delays are enabled, enforce a mimimum delay (hardcoded
to 10ms for now).

This can help even on short RTTs, and it is especially useful in the
presence of transparent TCP accelerator proxies that masquerade as the
TCP remote peer and are located right before a high-latency path, such
as a GEO satellite link.

This should be enhanced in the future to be an extended command line
parameter, and added as an optional measurement profile parameter.