[MDC-1410] simet-api: switch SIMETBox to N0 virtual labels
Change the SIMETBox version of simet_create_vlabel.sh to use N0 virtual
labels.

Note that a previous virtual label would have to be removed before this
will have any effect.

For auto-generated virtual labels, we retain the old code, and use it in
the vlabel-was-foreign codepath.  However, usually this codepath is not
needed because the MAC address checking done by the simetbox'
simet_detect_cloned_id() hook for simet_register-ma.sh takes precedence.