[MDC-1453] twamp: NTP era 1 support (fix Y2036 bug)
Handle the NTPv4 wire packet timestamp rollover on year 2036, also known
as the Y2036 bug, Y2K36 bug, or NTP epoch rollover issue.

Note: this is for the NTP rollover, it has nothing to do with the Y2038
bug which is related to UNIX timestamps going negative on platforms
where time_t is a 32-bit signed integer.

Consider that any NTP unsigned 32-bit timestamp that results in an year
before 2010 is in era 1 and shift it to the future by 2^32 seconds.
twampc now will translate NTPv4 wire timestamps into dates from 2010 to
2146 (on 64-bit time_t platforms).

We *could* use the system real time to select the NTP era instead, but
this doesn't look useful at the moment.