[MDC-1461] inetup: break endless loop on fscanf() failure
fscanf() will not set errno to zero when it returns EOF due to an
end-of-file condition.  That means one has to ensure errno is reset
between fscanf() calls or use feof() to differentiate the two reasons it
might return EOF.

inetup was failing to do so, and would loop forever at 100% CPU usage if
the agent-id or agent token files were empty at the time it got a
SIGHUP.

Limit the number of retries on EINTR, reset both the stream error flag
and errno between loops, AND use feof() as well for good measure.