[MDC-931] simet-runner: limit parallel traceroutes
Limit the number of parallel traceroutes depending on available memory,
or disable parallel execution of traceroutes.

Use --no-parallel to disable, and --limit-parallel-memory X (in KiB) to
change the default limit.  This can be used for other paralellization as
well.

By default, run traceroutes in parallel as long as there is at least
5MiB worth of available memory.