[MDC-1483] simet_api: simetbox: use ubus to get hostname, description
OpenWRT 23+ (and possibly OpenWRT 22) does not allow a non-privileged
user to retrieve the system hostname and description using the "uci"
utility, due to filesystem permissions.

This results in the lack of hostname and description metadata when
simet_register-ma.sh is called from the nicbr-simet non-privileged user.

Use ubus to call into the simet_ma rpc helper (which runs as root) to
retrieve the required information, just like we already do for engine
configuration in simetbox_lmap-fetch-schedule.sh.  Fallback to calling
the "uci" utility directly.