[MDC-1364] deb: fix simet-ma.cron.weekly fallback package list
Due to the lack of pipefail mode, we might not use the fallback package
list if the dpkg-query command fails, as the pipeline might return an
exit status of zero regardless.

Fix that by using the fallback package list if we don't get a list of
packages out of the dpkg-query | sed pipeline, regardless of pipeline
exit status.

Fixes: commit b30417084304402f05723af2579ad8f5aabbc6f0
       ([MDC-1364] deb: simet-ma.cron.weekly: update all simet* packages)