[MDC-1500] simet-api: simet_register-ma.sh: add --daily cli option
Add a --daily CLI option (which is also implied by the already existing
--boot CLI option) that rate-limits simet_register_ma to run once every
24H.

Either --boot or --daily *must* be used when calling simet_register_ma
from cron or other periodic script in a fixed schedule that has a period
shorter than 24H *or* which has *not* been randomized per-unit so that
there is no group behavior possible.

E.g. it must be used in cron.daily/weekly/monthly and typical crontab
invocations (that *includes* anacrontab without a large RANDOM_DELAY).