[MDC-1500] simet-api: simet_register_ma: harden timestamp handling
Reset (set to zero) the lastrun timestamp if it is missing, negative or
non-numeric, even if we don't otherwise update it to the current time.

If the system clock moved backwards, the lastrun timestamp could be in
the future.  In that case, reset it and allow simet_register_ma to run.