[MDC-1506] simet-api: simet_register_ma: rework API response handling
Refactor the response status handling and error handling when doing
SIMET2 AR API calls, improve error messages, and improve the
rate-limiting when something fails.

The rate-limiter timestamp is to be updated on sucessful runs, as well
as on situations where the client is buggy or running outdated software
(status 400 and 404 in particular, status 4xx in general, and when
deny-listed).

The rate-limiter timestamp is not to be updated on partially-successful
or failed runs, where relevant AR requests ultimately failed (*after*
retries where attempted) due to connection issues (non-zero return
status from curl) or due to server-side issues: status 500, 501, and 503
(status 5xx in general).  We also don't update it on situations where a
retry a short time later is not harmful: status 408 and 429.

If deny-listed by status 410 in POST /v2/agents, the rate-limiter is
*always* updated.

The rate-limiter timestamp is also updated on status 401 and 403
(invalid or expired credentials) on endpoints that don't trigger
immediate self-healing.

The rate-limiter timestamp is reset on package/firmware upgrades, and
reboot (to attempt to self-heal issues caused by broken system clock
handling in the plaform, and also unset system clocks).

The error paths have been revised, and when hooks fail it will not cause
a silent error exit anymore.  Some error and information log messages
changed.  In particular, when hooks are involved, extra (usually
redundant) error messages might now be generated: their value exists
when an override hook fails without proper error message feedback.