[MDC-1502] simet-api, deb: move simet_register_ma lastrun data to AGENT_TOKEN_LOCK
Now that simet_register_ma has been fixed to use AGENT_TOKEN_LOCK as its
main lock, move the simet_register_ma last-run information from
AGENT_LOCK to AGENT_TOKEN_LOCK, and leave AGENT_LOCK free to use, e.g.,
for measurement rate-limiting should we implement that in the future.

X-Vendor-Notice: Ensure the contents of the directory defined in
SIMET_LOCK_DIR are *NOT* persisted across reboots, power-cycle, or
firmware updates.  This is not a new rule.

X-Vendor-Notice: the file defined in AGENT_TOKEN_LOCK must be truncated
to size 0 on package updates.  Previously, this requirement existed for
AGENT_LOCK, instead.  It is not a problem to *also* truncate the file
defined in AGENT_LOCK (in fact, it is better to always truncate both).